Add feature flag for superadmin access.
