+++ coral/src/app/context-provider/AuthProvider.tsx
@@ -5,6 +5,8 @@ import { BasePage } from "src/app/layout/page/BasePage";
 import { Box, Icon } from "@aivenio/aquarium";
 import loading from "@aivenio/aquarium/icons/loading";
 import { NoCoralAccessSuperadmin } from "src/app/components/NoCoralAccessSuperadmin";
+import useFeatureFlag from "src/services/feature-flags/hook/useFeatureFlag";
+import { FeatureFlag } from "src/services/feature-flags/types";
 
 /** We don't do Authentication on Corals side
  * at the moment, so we only have a AuthUser
@@ -56,9 +58,18 @@ const AuthProvider = ({ children }: { children: ReactNode }) => {
     getAuth
   );
 
+  const superAdminAccessCoralEnabled = useFeatureFlag(
+    FeatureFlag.FEATURE_FLAG_SUPER_ADMIN_ACCESS_CORAL
+  );
+
   // SUPERADMIN does not have access to coral, so we show a reduced page with
   // information about that and nothing else.
-  if (!isLoading && authUser && isSuperAdmin(authUser)) {
+  if (
+    !isLoading &&
+    authUser &&
+    isSuperAdmin(authUser) &&
+    !superAdminAccessCoralEnabled
+  ) {
     return (
       <BasePage headerContent={<></>} content={<NoCoralAccessSuperadmin />} />
     );
