+++ coral/src/services/feature-flags/types.ts
@@ -1,5 +1,6 @@
 enum FeatureFlag {
   FEATURE_FLAG_ADD_CLUSTER = "FEATURE_FLAG_ADD_CLUSTER",
+  FEATURE_FLAG_SUPER_ADMIN_ACCESS_CORAL = "FEATURE_FLAG_SUPER_ADMIN_ACCESS_CORAL",
 }
 
 export { FeatureFlag };
